requirements.txt
# Core Dependencies
anthropic>=0.40.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Database
supabase>=2.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pgvector>=0.2.0

# WeChat Processing
pymupdf>=1.23.0
pdfplumber>=0.10.0
paddleocr>=2.7.0
pytesseract>=0.3.10
Pillow>=10.0.0
wechat-export-parser>=0.1.0

# External Sources
tweepy>=4.14.0
praw>=7.7.0
PyGithub>=2.1.0
newspaper3k>=0.2.8
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0

# Financial Data
yfinance>=0.2.0
akshare>=1.12.0
tushare>=1.4.0
openbb>=4.0.0
pandas>=2.0.0
numpy>=1.24.0

# NLP & Embeddings
openai>=1.0.0
sentence-transformers>=2.2.0
spacy>=3.7.0
jieba>=0.42.1

# Trading & Backtesting
vectorbt>=0.26.0
backtrader>=1.9.0
ta-lib>=0.4.0

# Memory & Context
chromadb>=0.4.0
langchain>=0.1.0
minecontext>=0.1.0

# Dashboard
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
plotly>=5.17.0
dash>=2.14.0

# Utilities
loguru>=0.7.0
tenacity>=8.2.0
asyncio>=3.4.3
aiohttp>=3.9.0
redis>=5.0.0
celery>=5.3.0
